move Events.js to root folder because it makes more sense
