+++ src/App.test.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import { BrowserRouter as Router } from "react-router-dom";
 import ROUTES from "./Router.js";
-import EVENTS from "./views/Home/Events.js";
+import EVENTS from "./Events.js";
 import fs from "fs";
 
 expect.extend({
