+++ src/views/Home/Home.js
@@ -5,7 +5,7 @@ import ArticleList from "../../components/ArticleList/ArticleList.js";
 import Contributions from "../../components/Contributions/Contributions.js";
 import DocumentTitle from "../../components/DocumentTitle/DocumentTitle.js";
 import EventBanner from "../../components/EventBanner/EventBanner.jsx";
-import EVENTS from "./Events.js";
+import EVENTS from "../../Events.js";
 import Logo from "../../components/Logo/Logo.jsx";
 import * as SEC from "../../components/SEC/lib.js";
